Update POM, fixes small worldguard checking error
Fixes bug as seen in https://pastebin.com/XFiGszw6 and in the discord bug reports. MythicMobs doesn't need to be an included JAR anymore, updated their Maven repo. WorldGuardRegionHook will abort if the world is null instead of trying to one-line call a giant method. Fixes some PhatLoots tasks that are still running trying to check on a world that has been deleted or something else.